[Chore #165263819]: refactor some server codes
