+++ api/package.json
@@ -4,7 +4,7 @@
   "description": "Nodejs server-side API development for Banka App",
   "main": "index.js",
   "scripts": {
-    "test": "NODE_ENV=TEST nyc mocha ./test --require @babel/register -w --exit || true",
+    "test": "nyc mocha ./test --require @babel/register --watch --exit || true",
     "dev-start": "nodemon --exec babel-node index.js"
   },
   "repository": {
