+++ api/test/user.spec.js
@@ -10,34 +10,29 @@ const should = chai.should();
 describe('Mocha test for User Controller', () => {
   describe('Mocha test for user signup route', () => {
     const signupUrl = '/api/v1/auth/signup';
-    it(
-      'should register a new user when all the parameters are given',
-      (done) => {
-        chai.request(app)
-          .post(signupUrl)
-          .send({
-            id: 1,
-            email: 'frank@email.com',
-            firstName: 'Frank',
-            lastName: 'Obi',
-            password: '12345',
-            type: 'client',
-            isAdmin: false
-          })
-          .end((error, response) => {
-            expect(response.body).to.be.an('object');
-            expect(response).to.have.status(201);
-            expect(response.body.status).to.equal(201);
-            expect(response.body.data).to.be.a('object');
-            expect(response.body.data).to.have.property('id');
-            expect(response.body.data).to.have.property('firstName');
-            expect(response.body.data).to.have.property('lastName');
-            expect(response.body.data).to.have.property('email');
-            expect(response.body.data.email).to.equal('frank@email.com');
-            done();
-          });
-      },
-    );
+    it('should register a new user when all the parameters are given', (done) => {
+      chai.request(app)
+        .post(signupUrl)
+        .send({
+          email: 'frank@email.com',
+          firstName: 'Frank',
+          lastName: 'Obi',
+          password: '12345',
+        })
+        .end((error, response) => {
+          expect(response.body).to.be.an('object');
+          expect(response.status).to.equal(201);
+          expect(response.body.status).to.equal(201);
+          expect(response.data).to.be.an('object');
+          expect(response.body.data).to.have.property('id');
+          expect(response.body.data).to.have.property('firstName');
+          expect(response.body.data).to.have.property('lastName');
+          expect(response.body.data).to.have.property('email');
+          expect(response.body.data.email).to.equal('frank@email.com');
+          console.log(response.body.data);
+          done();
+        });
+    });
 
     it('should not register a user when the email is missing', (done) => {
       chai.request(app)
